LDEV-3668:Fix for toggle subscription to receive email notifications by email.

+++ lams_common/src/java/org/lamsfoundation/lams/events/EventNotificationService.java
@@ -106,7 +106,7 @@ public class EventNotificationService implements IEventNotificationService {
 	Event event = eventDAO.getEvent(scope, name, eventSessionId);
 	if (event != null) {
 	    for (Subscription subscription : event.getSubscriptions()) {
-		if (subscription.getUserId().equals(userId)) {
+		if (subscription.getUserId().equals(Integer.valueOf(userId.intValue()))) {
 		    return true;
 		}
 	    }
